backend/lambda/requirements.txt
psycopg2-binary
bcrypt
PyJWT
boto3
